+++ Samsam/ViewController/HistoryView/WorkingHistoryView/WorkingHistoryViewController.swift
@@ -9,54 +9,22 @@ import UIKit
 
 class WorkingHistoryViewController: UIViewController {
     
-    // MARK: - Property
-    
-    var roomID: Int?
-    var views: [UIView?] = []
-    var index = 0
-    
     // MARK: - View
-    
-    private let segmentedControl: UISegmentedControl = {
-        $0.tintColor = .black
-        $0.backgroundColor = .clear
-        $0.selectedSegmentIndex = 0
-        $0.addTarget(self, action: #selector(tapSegmentButton), for: .valueChanged)
-        return $0
-    }(UISegmentedControl(items: ["시공내역", "문의내역"]))
-    
-    private let writingButton: UIButton = {
-        $0.backgroundColor = AppColor.campanulaBlue
-        $0.setTitle("시공상황 작성하기", for: .normal)
-        $0.titleLabel?.font = UIFont.systemFont(ofSize: 18, weight: .semibold)
-        $0.setTitleColor(.white, for: .normal)
-        $0.layer.cornerRadius = 16
-        return $0
-    }(UIButton())
     
     private let workingHistoryView: UICollectionView = {
         return $0
     }(UICollectionView(
         frame: .zero, collectionViewLayout: UICollectionViewFlowLayout()))
     
-    private let inquiryHistoryView: UIView = {
-        $0.backgroundColor = .black
-        return $0
-    }(UIView())
-    
     // MARK: - LifeCycle
     
     override func viewDidLoad() {
         super.viewDidLoad()
-        setNavigationBar()
-        
         attribute()
         layout()
     }
     
     override func viewWillAppear(_ animated: Bool) {
-        coreDataManager.loadOneRoomData(roomID: roomID!)
-        coreDataManager.loadPostingData(roomID: roomID!)
         workingHistoryView.reloadData()
     }
     
@@ -64,7 +32,6 @@ class WorkingHistoryViewController: UIViewController {
     
     private func attribute() {
         view.backgroundColor = .white
-        setSegmentedControl()
         
         workingHistoryView.delegate = self
         workingHistoryView.dataSource = self
@@ -73,124 +40,17 @@ class WorkingHistoryViewController: UIViewController {
         workingHistoryView.register(WorkingHistoryViewContentHeader.self, forSupplementaryViewOfKind: UICollectionView.elementKindSectionHeader, withReuseIdentifier: WorkingHistoryViewContentHeader.identifier)
         workingHistoryView.register(WorkingHistoryViewTopCell.self, forCellWithReuseIdentifier: WorkingHistoryViewTopCell.identifier)
         workingHistoryView.register(WorkingHistoryViewContentCell.self, forCellWithReuseIdentifier: WorkingHistoryViewContentCell.identifier)
-        
-        writingButton.addTarget(self, action: #selector(tapWritingButton), for: .touchDown)
     }
     
     private func layout() {
-        view.addSubview(segmentedControl)
-        view.addSubview(inquiryHistoryView)
         view.addSubview(workingHistoryView)
-        view.addSubview(writingButton)
-        
-        segmentedControl.anchor(
-            top: view.safeAreaLayoutGuide.topAnchor,
-            left: view.safeAreaLayoutGuide.leftAnchor,
-            right: view.safeAreaLayoutGuide.rightAnchor
-        )
-        
-        inquiryHistoryView.anchor(
-            top: segmentedControl.bottomAnchor,
-            left: view.safeAreaLayoutGuide.leftAnchor,
-            bottom: view.bottomAnchor,
-            right: view.safeAreaLayoutGuide.rightAnchor,
-            paddingTop: 16
-        )
         
         workingHistoryView.anchor(
-            top: segmentedControl.bottomAnchor,
-            left: view.safeAreaLayoutGuide.leftAnchor,
+            top: view.topAnchor,
+            left: view.leftAnchor,
             bottom: view.bottomAnchor,
-            right: view.safeAreaLayoutGuide.rightAnchor,
-            paddingTop: 16
+            right: view.rightAnchor
         )
-
-        writingButton.anchor(
-            left: view.safeAreaLayoutGuide.leftAnchor,
-            bottom: view.safeAreaLayoutGuide.bottomAnchor,
-            right: view.safeAreaLayoutGuide.rightAnchor,
-            paddingLeft: 16,
-            paddingRight: 16,
-            height: 50
-        )
-    }
-    
-    // MARK: - Method
-    
-    private func setNavigationBar() {
-        navigationItem.title = coreDataManager.oneRoom?.clientName
-        navigationController?.navigationBar.prefersLargeTitles = false
-        let rightBarButtonItem = UIBarButtonItem(image: UIImage(systemName: "gearshape"), style: .plain, target: self, action: #selector(tapSettingButton))
-        rightBarButtonItem.tintColor = .black
-        navigationItem.rightBarButtonItem = rightBarButtonItem
-    }
-
-    @objc func tapWritingButton() {
-        let postingCategoryViewController = PostingCategoryViewController()
-        postingCategoryViewController.roomID = roomID
-        let navigationController = UINavigationController(rootViewController: postingCategoryViewController)
-        navigationController.modalPresentationStyle = .fullScreen
-        present(navigationController, animated:  true, completion: nil)
-    }
-    
-    @objc func tapSettingButton() {
-        let settingViewController = ViewController()
-        navigationController?.pushViewController(settingViewController, animated: true)
-    }
-    
-    @objc func setSegmentedControl(){
-        views = [workingHistoryView, inquiryHistoryView]
-
-        let swipeLeft = UISwipeGestureRecognizer(target: self, action: #selector(handleGesture(gesture:)))
-        swipeLeft.direction = .left
-        self.view.addGestureRecognizer(swipeLeft)
-
-        let swipeRight = UISwipeGestureRecognizer(target: self, action: #selector(handleGesture(gesture:)))
-        swipeRight.direction = .right
-        self.view.addGestureRecognizer(swipeRight)
-    }
-    
-    @objc func handleGesture(gesture: UISwipeGestureRecognizer) {
-        if index >= 0 && index < views.count {
-            if gesture.direction == UISwipeGestureRecognizer.Direction.right {
-                if index != 0 {
-                    index -= 1
-                    self.view.bringSubviewToFront(views[index]!)
-                }
-            }
-            if gesture.direction == UISwipeGestureRecognizer.Direction.left {
-                if index != views.count - 1 {
-                    index += 1
-                    self.view.bringSubviewToFront(views[index]!)
-                }
-            }
-            self.view.bringSubviewToFront(writingButton)
-            segmentedControl.selectedSegmentIndex = index
-        }
-    }
-    
-    @objc func tapSegmentButton() {
-        switch segmentedControl.selectedSegmentIndex {
-        case 0:
-            self.view.bringSubviewToFront(workingHistoryView)
-            self.view.bringSubviewToFront(writingButton)
-        case 1:
-            self.view.bringSubviewToFront(inquiryHistoryView)
-            self.view.bringSubviewToFront(writingButton)
-        default:
-            break
-        }
-    }
-    
-    func moveToFrame(contentOffset : CGFloat) {
-
-        let frame: CGRect = CGRect(
-            x: contentOffset,
-            y: self.workingHistoryView.contentOffset.y,
-            width: self.workingHistoryView.frame.width,
-            height: self.workingHistoryView.frame.height
-        )
-        self.workingHistoryView.scrollRectToVisible(frame, animated: true)
     }
 }
 
@@ -206,7 +66,7 @@ extension WorkingHistoryViewController: UICollectionViewDataSource, UICollection
         if section == 0 {
             return CGSize(width: UIScreen.main.bounds.width, height: 100)
         }
-        return CGSize(width: UIScreen.main.bounds.width, height: 70)
+        return CGSize(width: UIScreen.main.bounds.width, height: 60)
     }
     
     func collectionView(_ collectionView: UICollectionView, viewForSupplementaryElementOfKind kind: String, at indexPath: IndexPath) -> UICollectionReusableView {
@@ -258,11 +118,12 @@ extension WorkingHistoryViewController: UICollectionViewDataSource, UICollection
         
         if indexPath.section == 0 {
             let width = UIScreen.main.bounds.width
-            let cellHeight = 20
+            let cellHeight = 30
             return CGSize(width: Int(width), height: cellHeight)
         } else {
             let width = UIScreen.main.bounds.width - 32
             let cellHeight = width / 4 * 3 + 30
+            
             return CGSize(width: Int(width), height: Int(cellHeight))
         }
     }
